+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/symboltable/TypeSet.java
@@ -168,7 +168,7 @@ public class TypeSet {
                         String importPkg = importStmt.substring(0, importStmt.indexOf('*') - 1);
                         return pmdClassLoader.loadClass(importPkg + '.' + name);
                     } catch (ClassNotFoundException cnfe) {
-                    } catch (NoClassDefFoundError ignored) {
+                        // ignored as the class could be imported with the next on demand import...
                     }
                 }
             }
@@ -198,7 +198,7 @@ public class TypeSet {
         @Override
         public Class<?> resolve(String name) throws ClassNotFoundException {
             if (!primitiveTypes.containsKey(name)) {
-                throw new ClassNotFoundException();
+                throw new ClassNotFoundException(name);
             }
             return primitiveTypes.get(name);
         }
@@ -213,7 +213,7 @@ public class TypeSet {
             if (name.equals("void")) {
                 return void.class;
             }
-            throw new ClassNotFoundException();
+            throw new ClassNotFoundException(name);
         }
     }
 
@@ -280,7 +280,7 @@ public class TypeSet {
             try {
                 return resolver.resolve(name);
             } catch (ClassNotFoundException cnfe) {
-            } catch (NoClassDefFoundError ignored) {
+                // ignored, maybe another resolver will find the class
             }
         }